Add configuration for max negociable block size to server
Add --max-blksize option to the tftpy_server in order to control
the max negociable block size.

This can be useful for reducing timeouts when client block size
requests are too large. For instance when the network is known
to be too loaded for correctly managing block sizes which cause
IP fragmentation.